Set path in .xprofile instead of .zshrc
xprofile is shell agnostic while zshrc is zsh specific. Apparently i3
uses sh to exec programs (at least that is what I read) so my custom
PATH is not set in these programs. This is a problem for emacs since I
launch it using rofi which is set to a keybinding in the i3 config.

Signed-off-by: Steven Erenst <stevenerenst@gmail.com>